next try to fix github workflow
GitHubs JAVA_HOME_V_A is a bit of a rabbit hole. It's difficult to work
with and the architecture tag is inconsitent. For our purposes with only
X64 and arm64 we will use a cheap hack I found in some github discussion
about this.

While there, fix some indenting.